refactor(adkg/feldman): move poly eval to helpers
